requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.31.0
openai==0.27.10
